fix: Support multiple namespace
